Bufixing + refactoring
