+++ src/store/hooks/useImages.ts
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import murmur from 'murmurhash';
 import { LoadedIIIFImage, LoadedImage } from '@/model';
 import { CozyCanvas } from '@/utils/cozy-iiif';
-import { fetchManifest } from '@/utils/iiif/utils/fetchManifest';
+import { fetchManifest } from '@/utils/iiif/fetchManifest';
 import { useStore } from './useStore';
 
 export const useImages = (
