+++ src/utils/iiif/fetchManifest.ts
