+++ src/utils/iiif/hooks/useIIIFResource.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { useStore } from '@/store';
-import { fetchManifest } from '../utils/fetchManifest';
+import { fetchManifest } from '../fetchManifest';
 import { CozyManifest } from '@/utils/cozy-iiif';
 
 export const useIIIFResource = (id: string) => {
