--- src/utils/iiif/lib/openseadragon/index.ts
@@ -1,1 +1,0 @@
-export * from './tilesets';
