--- src/utils/iiif/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './parseIIIFId';
-export * from './getOSDTilesets';
